requirements.txt
asgiref==3.8.1
boto3==1.34.145
botocore==1.34.145
cffi==1.16.0
cryptography==42.0.8
distlib==0.3.6
Django==4.1.4
django-cors-headers==4.3.1
django-storages==1.14.4
djangorestframework==3.15.1
filelock==3.8.2
jmespath==1.0.1
pillow==10.4.0
platformdirs==2.6.0
psycopg-binary==3.2.1
psycopg2-binary==2.9.9
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.17.1
